pyproject.toml
[tool.poetry]
name = "checkin-bot"
version = "0.1.0"
description = ""
authors = ["Shakhlyn <shakhlyn.sh.du@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py-cord = "^2.6.1"
python-dotenv = "^1.1.0"
apscheduler = "^3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev= 'checkin_bot.bot:main'